Swaggerized Veteran Model and Controller
All models and controllers will need to be swaggerized. See
https://swagger.io for details.  For the remaining models and
controllers, use the Veteran model and controller as an example.

For the models, use the FSS data dictionary to determine what goes
there.  some types are explanatory, but here's ones that don't
explicitly match: tinyint(1) is boolean, varchar(255) is string, and
timestamp is integer.

For the controllers, see routes.php to determine the method of each
operation.  Essentially, any delete methods are DELETE, update methods
are PUT, create methods are POST, and any read methods are GET.